AS-5077

+++ src/main/java/com/asg/shipchandling/salesquotationsch/dto/SalesQuotationSchHdrDto.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 import java.util.List;
 
@@ -23,7 +24,7 @@ public class SalesQuotationSchHdrDto {
     private Boolean newAddressYN;
 
     private String currencyCode;
-    private Long currencyRate;
+    private BigDecimal currencyRate;
     private String quotationStatus;
     private Long salesmanPoid;
     private Timestamp validityFromDate;
