+++ src/main/java/com/asg/shipchandling/salesquotationsch/dto/request/UpdateSalesQuotationSchRequest.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 import java.util.List;
 
@@ -28,7 +29,7 @@ public class UpdateSalesQuotationSchRequest {
     @Size(max = 20, message = "Currency code must not exceed 20 characters")
     private String currencyCode;
 
-    private Long currencyRate;
+    private BigDecimal currencyRate;
 
     @Size(max = 20, message = "Quotation status must not exceed 20 characters")
     private String quotationStatus;
